cleaning and renaming a bunch of operator in order to get consistent naming.
