+++ src/dynamic_graph/sot/core/math_small_entities.py
@@ -11,10 +11,10 @@ from .operator import (
     HomoToTwist, Inverse_of_matrix, Inverse_of_matrixHomo,
     Inverse_of_matrixrotation, Inverse_of_matrixtwist, Inverse_of_unitquat,
     MatrixDiagonal, MatrixHomoToPose, MatrixHomoToPoseQuaternion,
-    PoseQuatToMatrixHomo, MatrixHomoToPoseRollPitchYaw, MatrixHomoToPoseUTheta,
-    MatrixToHomo, MatrixToQuaternion, MatrixToRPY, MatrixToUTheta,
-    MatrixHomoToSE3Vector, SE3VectorToMatrixHomo, MatrixTranspose,
-    Multiply_double_vector, Multiply_matrix_vector,
+    PoseQuaternionToMatrixHomo, MatrixHomoToPoseRollPitchYaw,
+    MatrixHomoToPoseUTheta, MatrixToHomo, MatrixToQuaternion, MatrixToRPY,
+    MatrixToUTheta, MatrixHomoToSE3Vector, SE3VectorToMatrixHomo,
+    MatrixTranspose, Multiply_double_vector, Multiply_matrix_vector,
     Multiply_matrixTwist_vector, Multiply_matrixHomo_vector,
     Multiply_of_double, Multiply_of_matrix, Multiply_of_matrixHomo,
     Multiply_of_matrixrotation, Multiply_of_matrixtwist,
