+++ src/dynamic_graph/sot/core/meta_task_visual_point.py
@@ -3,7 +3,7 @@ from dynamic_graph.sot.core.feature_visual_point import FeatureVisualPoint
 from dynamic_graph.sot.core.gain_adaptive import GainAdaptive
 from dynamic_graph.sot.core.meta_tasks import setGain
 from dynamic_graph.sot.core.op_point_modifier import OpPointModifier
-from dynamic_graph.sot.core.sot import Task
+from dynamic_graph.sot.core.task import Task
 from dynamic_graph.sot.core.visual_point_projecter import VisualPointProjecter
